% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_plot_trajectory_function.R
\name{monocle_theme_opts}
\alias{monocle_theme_opts}
\title{Loads theme}
\usage{
monocle_theme_opts()
}
\description{
Accepts a subset of a CellDataSet and an attribute to group cells by,
and produces one or more ggplot2 objects that plots the level of expression for
each group of cells.
}
