% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_plot_trajectory_function.R
\name{my_plot_cell_trajectory}
\alias{my_plot_cell_trajectory}
\title{Plots cell trajectory}
\usage{
my_plot_cell_trajectory(cds, x = 1, y = 2, color_by = "State",
  show_tree = TRUE, show_backbone = TRUE, backbone_color = "black",
  markers = NULL, markers_linear = FALSE, show_cell_names = FALSE,
  show_state_number = FALSE, cell_size = 1.5, cell_link_size = 0.75,
  cell_name_size = 2, state_number_size = 2.9, show_branch_points = TRUE,
  point_colors = c("#cb4bbe", "lightskyblue", "grey37", "#53ba48", "moccasin",
  "#dcca44", "#502d70", "#afd671", "#cb4471", "#69da99", "#d44732", "#6fd3cf",
  "#5d2539", "#cfd0a0", "blue", "#d2883b", "#6793c0", "#898938", "#c98ac2",
  "yellow", "#c4c0cc", "#7d3d23", "#00a5ff", "#d68d7a", "#a2c1a3"), theta = 0)
}
\description{
Accepts a subset of a CellDataSet and an attribute to group cells by,
and produces one or more ggplot2 objects that plots the level of expression for
each group of cells.
}
