% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_plot_trajectory_function.R
\name{my_plot_rho_delta}
\alias{my_plot_rho_delta}
\title{Plots the decision map of density clusters .}
\usage{
my_plot_rho_delta(cds, rho_threshold = NULL, delta_threshold = NULL)
}
\arguments{
\item{cds}{CellDataSet for the experiment after running clusterCells_Density_Peak}

\item{rho_threshold}{The threshold of local density (rho) used to select the density peaks for plotting}

\item{delta_threshold}{The threshold of local distance (delta) used to select the density peaks for plotting}
}
\description{
Plots the decision map of density clusters .
}
\examples{
\dontrun{
library(HSMMSingleCell)
HSMM <- load_HSMM()
plot_rho_delta(HSMM)
}
}
